Design nav-bar
